+++ Controls-demo/Spoiler/Header/headerTemplate.less
