+++ Controls-demo/Spoiler/HeaderRight/Index.ts
@@ -1,0 +1,9 @@
+import {Control, IControlOptions, TemplateFunction} from 'UI/Base';
+import controlTemplate = require('wml!Controls-demo/Spoiler/HeaderRight/HeaderRight');
+import 'css!Controls-demo/Controls-demo';
+
+class HeaderContentTemplate extends Control<IControlOptions> {
+    protected _template: TemplateFunction = controlTemplate;
+    static _theme: string[] = ['Controls/Classes'];
+}
+export default HeaderContentTemplate;
