--- Controls-demo/Spoiler/HeaderTemplate/template/Index.ts
@@ -1,12 +1,0 @@
-import {Control, IControlOptions, TemplateFunction} from 'UI/Base';
-import controlTemplate = require('wml!Controls-demo/Spoiler/HeaderTemplate/template/Template');
-import notifyHandler = require('Controls/Utils/tmplNotify');
-import 'css!Controls-demo/Spoiler/HeaderTemplate/headerTemplate';
-import 'css!Controls-demo/Controls-demo';
-
-class Template extends Control<IControlOptions> {
-    protected _template: TemplateFunction = controlTemplate;
-    protected _notifyHandler: Function = notifyHandler;
-    static _theme: string[] = ['Controls/Classes'];
-}
-export default Template;
