+++ Controls/_spoiler/View.ts
@@ -18,9 +18,9 @@ import Util from './Util';
 /**
  * @name Controls/_spoiler/IViewOptions#headerContentTemplate
  * @cfg {function|String} Контент, занимающий свободное пространство справа от заголовка. Если заголовка нет, то контент занимает все пространство шапки, в этом случае заголовок можно добавить вручную в любом месте.
+ * @demo Controls-demo/Spoiler/Header/Index
+ * @demo Controls-demo/Spoiler/HeaderRight/Index
  * @demo Controls-demo/Spoiler/HeadingLeft/Index
- * @demo Controls-demo/Spoiler/HeaderRightTemplate/Index
- * @demo Controls-demo/Spoiler/HeaderTemplate/Index
  */
